+++ documentation/source/CHANGELOG.md
@@ -36,8 +36,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - The group (--group, -g) filter
 - The import feature.
 - The authentication for Drupal now adds the following variables automatically: `${user.uid}, ${user.id}, ${user.name}, ${user.pass}` for the authenticated user. That means you can use these in subsequent tests, even if not authenticating said test.
-- `--show-request` and `--show-response`.
-- Request headers and body to display with `--show-request` or `--show-source`. If you do not want to see request and ONLY response, use `--show-response` instead of `--show-source`.
+- `--request` and `--response`.
+- Request headers and body to display with `--request` or `--show-source`. If you do not want to see request and ONLY response, use `--response` instead of `--show-source`.
 - --help and -h to printout CLI options.
 - A debug message if test is missing assertions.
 - Start and stop date and times.